Add files via upload
Problem Statement
1.	Explain the working and the differences between Maven, Gradle and SBT in detail.
Maven-
Maven is a build tool which are programs that automate the creation of executable applications from source code. In small projects, developers will often manually invoke the build process. This is not practical for larger projects, where it is very hard to keep track of what needs to be built, in what sequence and what dependencies there are in the building process. Using an automation tool allows the build process to be more consistent.
There are different types of Build tools available for different types of languages. For example
For java – Ant,Maven,Gradle.
For .NET framework – NAnt
For c# – MsBuild.
So Maven is a Build tool for java applications
Maven can do the following things-
-Build
-Document
-Dependency Management
-Distributions
-Test
-Report
-SCM’s
-Releases
Maven needs an XML file, that XML file is called as POM file stands for Project Object